Help dataset for plotting alignment scores on an axis

Dataset containing numeric values corresponding to alphabetic scores for easier plotting.

Usage

alignment_scores_values

Format

An object of class tbl_df (inherits from tbl, data.frame) with 6 rows and 5 columns.

Examples

alignment_scores_values
#> # A tibble: 6 × 5
#>   score_upper score_symbol category score_delta score_label
#>         <dbl> <chr>        <chr>          <dbl>       <dbl>
#> 1        16.5 E            score           16.5        8.25
#> 2        33.2 D            score           16.7       24.8 
#> 3        49.9 C            score           16.7       41.6 
#> 4        66.6 B            score           16.7       58.2 
#> 5        83.3 A            score           16.7       74.9 
#> 6       100   A+           score           16.7       91.6