Prepares data inputs needed for rendering the executive summary

Prepares data inputs needed for rendering the executive summary

Usage

prep_data_executive_summary(
  investor_name,
  portfolio_name,
  peer_group,
  start_year,
  scenario_source,
  scenario_selected,
  scenario_geography,
  equity_market,
  portfolio_allocation_method_equity,
  portfolio_allocation_method_bonds,
  green_techs,
  equity_results_portfolio,
  bonds_results_portfolio,
  peers_equity_results_aggregated,
  peers_bonds_results_aggregated,
  peers_equity_results_individual,
  peers_bonds_results_individual,
  indices_equity_results_portfolio,
  indices_bonds_results_portfolio,
  audit_file,
  emissions_portfolio,
  index_dir
)

Arguments

investor_name

Some description

portfolio_name

Some description

peer_group

Some description

start_year

Some description

scenario_source

Some description

scenario_selected

Some description

scenario_geography

Some description

equity_market

Some description

portfolio_allocation_method_equity

Some description

portfolio_allocation_method_bonds

Some description

green_techs

Some description

equity_results_portfolio

Some description

bonds_results_portfolio

Some description

peers_equity_results_aggregated

Some description

peers_bonds_results_aggregated

Some description

peers_equity_results_individual

Some description

peers_bonds_results_individual

Some description

indices_equity_results_portfolio

Some description

indices_bonds_results_portfolio

Some description

audit_file

Some description

emissions_portfolio

Some description

index_dir

Directory containing .rds files with index benchmarks

Value

data.frame