Renders executive summary

Renders executive summary

Usage

render_executive_summary(
  data,
  language,
  output_dir,
  exec_summary_dir,
  survey_dir,
  score_card_dir,
  analysis_inputs_dir,
  file_name = "template.Rmd",
  investor_name,
  portfolio_name,
  peer_group,
  total_portfolio,
  scenario_source = "WEO2023",
  scenario_selected = "NZE_2050",
  currency_exchange_value,
  log_dir
)

Arguments

data

List of data frames in the format returned by prep_data_executive_summary()

language

Character single, valid two letter language identifier in uppercase e.g. "EN"

output_dir

Character single, valid filepath to a directory where the output will be saved

exec_summary_dir

Character single, valid filepath to a directory that contains the template, e.g. system.file("extdata", "PA2022CH_en_exec_summary", package = "pacta.executive.summary")

survey_dir

Character single, valid filepath to a directory that contains the survey files for the user

score_card_dir

Character single, valid filepath to a directory that contains score card files for the user

analysis_inputs_dir

Character single, valid filepath to a PACTA analysis results directory that contains the SBTi data merged with financial data in the COP case

file_name

Character single, valid filename of the Rmd template file, e.g. "template.Rmd"

investor_name

Character single string specifying the investor name

portfolio_name

Character single string specifying the portfolio name

peer_group

Character single string specifying the peer group

total_portfolio

Data frame that contains the total portfolio as found in the standard PACTA processed inputs file "total_portfolio.rds"

scenario_source

Character single string specifying the selected scenario source, e.g. "GECO2023"

scenario_selected

Character single string specifying the selected scenario, e.g. "1.5C"

currency_exchange_value

Numeric single numeric value specifying the exchange rate from USD into the desired display currency, e.g. 1.03

log_dir

Character single, valid filepath to a directory that will contain the log file

Value

a pdf document written to output_dir